move PHPStanNodeScopeResolver under PHPStan namespace
