+++ packages/NodeTypeResolver/src/PHPStan/Scope/NodeScopeResolver.php
@@ -1,12 +1,12 @@
 <?php declare(strict_types=1);
 
-namespace Rector\NodeTypeResolver;
+namespace Rector\NodeTypeResolver\PHPStan\Scope;
 
 use PhpParser\Node;
 use PhpParser\Node\Stmt\Class_;
 use PhpParser\Node\Stmt\Interface_;
 use PhpParser\NodeVisitor\NameResolver;
-use PHPStan\Analyser\NodeScopeResolver;
+use PHPStan\Analyser\NodeScopeResolver as PHPStanNodeScopeResolver;
 use PHPStan\Analyser\Scope;
 use PHPStan\Broker\Broker;
 use Rector\Configuration\Option;
@@ -14,7 +14,6 @@ use Rector\Exception\ShouldNotHappenException;
 use Rector\FileSystem\FilesFinder;
 use Rector\Node\Attribute;
 use Rector\NodeTypeResolver\Configuration\CurrentFileProvider;
-use Rector\NodeTypeResolver\PHPStan\Scope\ScopeFactory;
 use Symplify\PackageBuilder\Parameter\ParameterProvider;
 use Symplify\PackageBuilder\Reflection\PrivatesAccessor;
 
@@ -22,12 +21,12 @@ use Symplify\PackageBuilder\Reflection\PrivatesAccessor;
  * @inspired by https://github.com/silverstripe/silverstripe-upgrader/blob/532182b23e854d02e0b27e68ebc394f436de0682/src/UpgradeRule/PHP/Visitor/PHPStanScopeVisitor.php
  * - https://github.com/silverstripe/silverstripe-upgrader/pull/57/commits/e5c7cfa166ad940d9d4ff69537d9f7608e992359#diff-5e0807bb3dc03d6a8d8b6ad049abd774
  */
-final class PHPStanNodeScopeResolver
+final class NodeScopeResolver
 {
     /**
-     * @var NodeScopeResolver
+     * @var PHPStanNodeScopeResolver
      */
-    private $nodeScopeResolver;
+    private $phpStanNodeScopeResolver;
 
     /**
      * @var CurrentFileProvider
@@ -59,14 +58,14 @@ final class PHPStanNodeScopeResolver
         ParameterProvider $parameterProvider,
         FilesFinder $filesFinder,
         ScopeFactory $scopeFactory,
-        NodeScopeResolver $nodeScopeResolver,
+        PHPStanNodeScopeResolver $phpStanNodeScopeResolver,
         Broker $broker
     ) {
         $this->currentFileProvider = $currentFileProvider;
         $this->parameterProvider = $parameterProvider;
         $this->filesFinder = $filesFinder;
         $this->scopeFactory = $scopeFactory;
-        $this->nodeScopeResolver = $nodeScopeResolver;
+        $this->phpStanNodeScopeResolver = $phpStanNodeScopeResolver;
         $this->broker = $broker;
     }
 
@@ -79,7 +78,7 @@ final class PHPStanNodeScopeResolver
         $this->ensureNameResolverWasRun($nodes);
         $this->setAnalysedFiles();
 
-        $this->nodeScopeResolver->processNodes(
+        $this->phpStanNodeScopeResolver->processNodes(
             $nodes,
             $this->scopeFactory->createFromFileInfo($this->currentFileProvider->getSplFileInfo()),
             function (Node $node, Scope $scope): void {
@@ -91,7 +90,7 @@ final class PHPStanNodeScopeResolver
                     } else {
                         // possibly anonymous class
                         $anonymousClassReflection = (new PrivatesAccessor())->getPrivateProperty(
-                            $this->nodeScopeResolver,
+                            $this->phpStanNodeScopeResolver,
                             'anonymousClassReflection'
                         );
 
@@ -118,7 +117,7 @@ final class PHPStanNodeScopeResolver
             $filePaths[] = $phpFileInfo->getRealPath();
         }
 
-        $this->nodeScopeResolver->setAnalysedFiles($filePaths);
+        $this->phpStanNodeScopeResolver->setAnalysedFiles($filePaths);
     }
 
     /**
